+++ src/ThreeDAnalyzer.Web/obj/Debug/net10.0/ThreeDAnalyzer.Web.AssemblyInfo.cs
@@ -13,7 +13,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("ThreeDAnalyzer.Web")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+c793d6b5cb62ed031d7f954406f8c59cf413de16")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+d88ce18f4ff24b3ce829522b495f3817978e8aec")]
 [assembly: System.Reflection.AssemblyProductAttribute("ThreeDAnalyzer.Web")]
 [assembly: System.Reflection.AssemblyTitleAttribute("ThreeDAnalyzer.Web")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]